Add CommandDetailSettingsPresenter
Add CommandDetailSettingsPresenter
