+++ ZIGSIMPlus/Models/AppSettingModel.swift
@@ -73,22 +73,6 @@ enum DepthType: Int {
     case DISPARITY = 1
 }
 
-public enum DetailSettingsKey: Int {
-    case arkitTrackingType = 0
-}
-
-public struct DetailSettingsSegmented {
-    var key: DetailSettingsKey
-    var label: String
-    var segments: [String]
-    var width: Int
-    var value: Int
-}
-
-public enum DetailSettingsData {
-    case segmented(DetailSettingsSegmented)
-}
-
 public class AppSettingModel {
     private init() {
         for command in Command.allCases {
@@ -171,14 +155,6 @@ public class AppSettingModel {
     public func isCameraEnabled() -> Bool {
         return AppSettingModel.shared.isActiveByCommand[.arkit]! ||
             AppSettingModel.shared.isActiveByCommand[.ndi]!
-    }
-
-    public func getCommandDetailSettings() -> [Command: [DetailSettingsData]] {
-        return [
-            .arkit: [
-                .segmented(DetailSettingsSegmented(key: .arkitTrackingType, label: "TRACKING TYPE", segments: ["DEVICE", "FACE", "MARKER"], width: 240, value: arkitTrackingType.rawValue))
-            ]
-        ]
     }
 }
